+++ releases/eucalyptus/3/wb/config/lms/docker_run_production.py
@@ -116,19 +116,6 @@ CELERY_QUEUES = config(
     formatter=json.loads,
 )
 
-# Setup alternate queues, to allow access to cross-process workers
-ALTERNATE_QUEUE_ENVS = config("ALTERNATE_WORKER_QUEUES", default="").split()
-ALTERNATE_QUEUES = [
-    DEFAULT_PRIORITY_QUEUE.replace("lms.", alternate + ".")
-    for alternate in ALTERNATE_QUEUE_ENVS
-]
-CELERY_QUEUES.update(
-    {
-        alternate: {}
-        for alternate in ALTERNATE_QUEUES
-        if alternate not in CELERY_QUEUES.keys()
-    }
-)
 CELERY_ROUTES = "lms.celery.Router"
 
 # Force accepted content to "json" only. If we also accept pickle-serialized